+++ ts/packages/shell/src/main/debug.ts
@@ -2,5 +2,7 @@
 // Licensed under the MIT License.
 
 import registerDebug from "debug";
+export const debugShellInit = registerDebug("typeagent:shell:init");
+export const debugShellCleanup = registerDebug("typeagent:shell:cleanup");
 export const debugShell = registerDebug("typeagent:shell");
 export const debugShellError = registerDebug("typeagent:shell:error");
